make_toggle_account_type_separate_component
