+++ packages/appstore/src/components/toggle-account-type/index.ts
@@ -1,0 +1,4 @@
+import ToggleAccountType from './toggle-account-type';
+import './toggle-account-type.scss';
+
+export default ToggleAccountType;
